+++ src/SlideDotNet/Models/SlideComponents/Chart/NoChart.cs
@@ -1,4 +1,5 @@
-﻿using SlideDotNet.Enums;
+﻿using SlideDotNet.Collections;
+using SlideDotNet.Enums;
 using SlideDotNet.Exceptions;
 
 namespace SlideDotNet.Models.SlideComponents.Chart
@@ -10,5 +11,11 @@ namespace SlideDotNet.Models.SlideComponents.Chart
         public string Title => throw new SlideDotNetException(ExceptionMessages.NoChart);
 
         public bool HasTitle => throw new SlideDotNetException(ExceptionMessages.NoChart);
+
+        public SeriesCollection SeriesCollection => throw new SlideDotNetException(ExceptionMessages.NoChart);
+
+        public CategoryCollection Categories => throw new SlideDotNetException(ExceptionMessages.NoChart);
+
+        public bool HasCategories => throw new SlideDotNetException(ExceptionMessages.NoChart);
     }
 }