+++ CHANGELOG.md
@@ -4,7 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Version 0.2.0 - Unreleased
+## Version 0.3.0 - 2020-03-16
+### Added
+- Added _ChartEx.SeriesCollection_ and  _ChartEx.Categories_ collections
+    ```
+    var pointValue = chart.SeriesCollection[0].PointValues[0];
+    var seriesType = chart.SeriesCollection[0].Type;
+    if (chart.HasCategories)
+    {
+        var category = chart.Categories[0];
+    }
+    ```
+
+## Version 0.2.0 - 2020-03-02
 ### Added
 - Added `SlideNumber` placeholder processing;
 - Added property `ShapeEx.Fill`.